+++ remote-client/src/test/java/org/opensearch/remote/metadata/client/impl/RemoteClusterIndicesClientTests.java
@@ -15,7 +15,7 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import org.opensearch.OpenSearchStatusException;
 import org.opensearch.action.DocWriteResponse;
 import org.opensearch.client.json.jackson.JacksonJsonpMapper;
-import org.opensearch.client.opensearch.OpenSearchClient;
+import org.opensearch.client.opensearch.OpenSearchAsyncClient;
 import org.opensearch.client.opensearch._types.ErrorCause;
 import org.opensearch.client.opensearch._types.ErrorResponse;
 import org.opensearch.client.opensearch._types.OpType;
@@ -116,7 +116,7 @@ public class RemoteClusterIndicesClientTests {
     );
 
     @Mock
-    private OpenSearchClient mockedOpenSearchClient;
+    private OpenSearchAsyncClient mockedOpenSearchAsyncClient;
     private SdkClient sdkClient;
 
     @Mock
@@ -128,14 +128,17 @@ public class RemoteClusterIndicesClientTests {
     public void setup() {
         MockitoAnnotations.openMocks(this);
 
-        when(mockedOpenSearchClient._transport()).thenReturn(transport);
+        when(mockedOpenSearchAsyncClient._transport()).thenReturn(transport);
         when(transport.jsonpMapper()).thenReturn(
             new JacksonJsonpMapper(
                 new ObjectMapper().setPropertyNamingStrategy(PropertyNamingStrategies.SNAKE_CASE)
                     .setSerializationInclusion(JsonInclude.Include.NON_NULL)
             )
         );
-        sdkClient = SdkClientFactory.wrapSdkClientDelegate(new RemoteClusterIndicesClient(mockedOpenSearchClient, TENANT_ID_FIELD), true);
+        sdkClient = SdkClientFactory.wrapSdkClientDelegate(
+            new RemoteClusterIndicesClient(mockedOpenSearchAsyncClient, TENANT_ID_FIELD),
+            true
+        );
         testDataObject = new TestDataObject("foo");
     }
 
@@ -162,7 +165,7 @@ public class RemoteClusterIndicesClientTests {
             .build();
 
         ArgumentCaptor<IndexRequest<?>> indexRequestCaptor = ArgumentCaptor.forClass(IndexRequest.class);
-        when(mockedOpenSearchClient.index(indexRequestCaptor.capture())).thenReturn(indexResponse);
+        when(mockedOpenSearchAsyncClient.index(indexRequestCaptor.capture())).thenReturn(CompletableFuture.completedFuture(indexResponse));
 
         PutDataObjectResponse response = sdkClient.putDataObjectAsync(putRequest, testThreadPool.executor(TEST_THREAD_POOL))
             .toCompletableFuture()
@@ -201,7 +204,7 @@ public class RemoteClusterIndicesClientTests {
             .build();
 
         ArgumentCaptor<IndexRequest<?>> indexRequestCaptor = ArgumentCaptor.forClass(IndexRequest.class);
-        when(mockedOpenSearchClient.index(indexRequestCaptor.capture())).thenReturn(indexResponse);
+        when(mockedOpenSearchAsyncClient.index(indexRequestCaptor.capture())).thenReturn(CompletableFuture.completedFuture(indexResponse));
 
         PutDataObjectResponse response = sdkClient.putDataObjectAsync(putRequest, testThreadPool.executor(TEST_THREAD_POOL))
             .toCompletableFuture()
@@ -231,7 +234,9 @@ public class RemoteClusterIndicesClientTests {
             .build();
 
         ArgumentCaptor<IndexRequest<?>> indexRequestCaptor = ArgumentCaptor.forClass(IndexRequest.class);
-        when(mockedOpenSearchClient.index(indexRequestCaptor.capture())).thenThrow(new IOException("test"));
+        when(mockedOpenSearchAsyncClient.index(indexRequestCaptor.capture())).thenReturn(
+            CompletableFuture.failedFuture(new IOException("test"))
+        );
 
         CompletableFuture<PutDataObjectResponse> future = sdkClient.putDataObjectAsync(
             putRequest,
@@ -255,7 +260,9 @@ public class RemoteClusterIndicesClientTests {
 
         ArgumentCaptor<GetRequest> getRequestCaptor = ArgumentCaptor.forClass(GetRequest.class);
         ArgumentCaptor<Class<Map>> mapClassCaptor = ArgumentCaptor.forClass(Class.class);
-        when(mockedOpenSearchClient.get(getRequestCaptor.capture(), mapClassCaptor.capture())).thenReturn((GetResponse<Map>) getResponse);
+        when(mockedOpenSearchAsyncClient.get(getRequestCaptor.capture(), mapClassCaptor.capture())).thenReturn(
+            CompletableFuture.completedFuture((GetResponse<Map>) getResponse)
+        );
 
         GetDataObjectResponse response = sdkClient.getDataObjectAsync(getRequest, testThreadPool.executor(TEST_THREAD_POOL))
             .toCompletableFuture()
@@ -284,7 +291,9 @@ public class RemoteClusterIndicesClientTests {
 
         ArgumentCaptor<GetRequest> getRequestCaptor = ArgumentCaptor.forClass(GetRequest.class);
         ArgumentCaptor<Class<Map>> mapClassCaptor = ArgumentCaptor.forClass(Class.class);
-        when(mockedOpenSearchClient.get(getRequestCaptor.capture(), mapClassCaptor.capture())).thenReturn((GetResponse<Map>) getResponse);
+        when(mockedOpenSearchAsyncClient.get(getRequestCaptor.capture(), mapClassCaptor.capture())).thenReturn(
+            CompletableFuture.completedFuture((GetResponse<Map>) getResponse)
+        );
 
         GetDataObjectResponse response = sdkClient.getDataObjectAsync(getRequest, testThreadPool.executor(TEST_THREAD_POOL))
             .toCompletableFuture()
@@ -303,7 +312,9 @@ public class RemoteClusterIndicesClientTests {
 
         ArgumentCaptor<GetRequest> getRequestCaptor = ArgumentCaptor.forClass(GetRequest.class);
         ArgumentCaptor<Class<Map>> mapClassCaptor = ArgumentCaptor.forClass(Class.class);
-        when(mockedOpenSearchClient.get(getRequestCaptor.capture(), mapClassCaptor.capture())).thenThrow(new IOException("test"));
+        when(mockedOpenSearchAsyncClient.get(getRequestCaptor.capture(), mapClassCaptor.capture())).thenReturn(
+            CompletableFuture.failedFuture(new IOException("test"))
+        );
 
         CompletableFuture<GetDataObjectResponse> future = sdkClient.getDataObjectAsync(
             getRequest,
@@ -334,7 +345,9 @@ public class RemoteClusterIndicesClientTests {
 
         @SuppressWarnings("unchecked")
         ArgumentCaptor<UpdateRequest<Map<String, Object>, ?>> updateRequestCaptor = ArgumentCaptor.forClass(UpdateRequest.class);
-        when(mockedOpenSearchClient.update(updateRequestCaptor.capture(), any())).thenReturn(updateResponse);
+        when(mockedOpenSearchAsyncClient.update(updateRequestCaptor.capture(), any())).thenReturn(
+            CompletableFuture.completedFuture(updateResponse)
+        );
 
         UpdateDataObjectResponse response = sdkClient.updateDataObjectAsync(updateRequest, testThreadPool.executor(TEST_THREAD_POOL))
             .toCompletableFuture()
@@ -375,7 +388,9 @@ public class RemoteClusterIndicesClientTests {
 
         @SuppressWarnings("unchecked")
         ArgumentCaptor<UpdateRequest<Map<String, Object>, ?>> updateRequestCaptor = ArgumentCaptor.forClass(UpdateRequest.class);
-        when(mockedOpenSearchClient.update(updateRequestCaptor.capture(), any())).thenReturn(updateResponse);
+        when(mockedOpenSearchAsyncClient.update(updateRequestCaptor.capture(), any())).thenReturn(
+            CompletableFuture.completedFuture(updateResponse)
+        );
 
         sdkClient.updateDataObjectAsync(updateRequest, testThreadPool.executor(TEST_THREAD_POOL)).toCompletableFuture().join();
 
@@ -407,7 +422,9 @@ public class RemoteClusterIndicesClientTests {
 
         @SuppressWarnings("unchecked")
         ArgumentCaptor<UpdateRequest<Map<String, Object>, ?>> updateRequestCaptor = ArgumentCaptor.forClass(UpdateRequest.class);
-        when(mockedOpenSearchClient.update(updateRequestCaptor.capture(), any())).thenReturn(updateResponse);
+        when(mockedOpenSearchAsyncClient.update(updateRequestCaptor.capture(), any())).thenReturn(
+            CompletableFuture.completedFuture(updateResponse)
+        );
 
         UpdateDataObjectResponse response = sdkClient.updateDataObjectAsync(updateRequest, testThreadPool.executor(TEST_THREAD_POOL))
             .toCompletableFuture()
@@ -436,7 +453,9 @@ public class RemoteClusterIndicesClientTests {
             .build();
 
         ArgumentCaptor<UpdateRequest<?, ?>> updateRequestCaptor = ArgumentCaptor.forClass(UpdateRequest.class);
-        when(mockedOpenSearchClient.update(updateRequestCaptor.capture(), any())).thenThrow(new IOException("test"));
+        when(mockedOpenSearchAsyncClient.update(updateRequestCaptor.capture(), any())).thenReturn(
+            CompletableFuture.failedFuture(new IOException("test"))
+        );
 
         CompletableFuture<UpdateDataObjectResponse> future = sdkClient.updateDataObjectAsync(
             updateRequest,
@@ -464,7 +483,9 @@ public class RemoteClusterIndicesClientTests {
                 .error(new ErrorCause.Builder().type("test").reason("test").build())
                 .build()
         );
-        when(mockedOpenSearchClient.update(updateRequestCaptor.capture(), any())).thenThrow(conflictException);
+        when(mockedOpenSearchAsyncClient.update(updateRequestCaptor.capture(), any())).thenReturn(
+            CompletableFuture.failedFuture(conflictException)
+        );
 
         CompletableFuture<UpdateDataObjectResponse> future = sdkClient.updateDataObjectAsync(
             updateRequest,
@@ -495,7 +516,9 @@ public class RemoteClusterIndicesClientTests {
             .build();
 
         ArgumentCaptor<DeleteRequest> deleteRequestCaptor = ArgumentCaptor.forClass(DeleteRequest.class);
-        when(mockedOpenSearchClient.delete(deleteRequestCaptor.capture())).thenReturn(deleteResponse);
+        when(mockedOpenSearchAsyncClient.delete(deleteRequestCaptor.capture())).thenReturn(
+            CompletableFuture.completedFuture(deleteResponse)
+        );
 
         DeleteDataObjectResponse response = sdkClient.deleteDataObjectAsync(deleteRequest, testThreadPool.executor(TEST_THREAD_POOL))
             .toCompletableFuture()
@@ -532,7 +555,9 @@ public class RemoteClusterIndicesClientTests {
             .build();
 
         ArgumentCaptor<DeleteRequest> deleteRequestCaptor = ArgumentCaptor.forClass(DeleteRequest.class);
-        when(mockedOpenSearchClient.delete(deleteRequestCaptor.capture())).thenReturn(deleteResponse);
+        when(mockedOpenSearchAsyncClient.delete(deleteRequestCaptor.capture())).thenReturn(
+            CompletableFuture.completedFuture(deleteResponse)
+        );
 
         DeleteDataObjectResponse response = sdkClient.deleteDataObjectAsync(deleteRequest, testThreadPool.executor(TEST_THREAD_POOL))
             .toCompletableFuture()
@@ -557,7 +582,9 @@ public class RemoteClusterIndicesClientTests {
             .build();
 
         ArgumentCaptor<DeleteRequest> deleteRequestCaptor = ArgumentCaptor.forClass(DeleteRequest.class);
-        when(mockedOpenSearchClient.delete(deleteRequestCaptor.capture())).thenThrow(new IOException("test"));
+        when(mockedOpenSearchAsyncClient.delete(deleteRequestCaptor.capture())).thenReturn(
+            CompletableFuture.failedFuture(new IOException("test"))
+        );
 
         CompletableFuture<DeleteDataObjectResponse> future = sdkClient.deleteDataObjectAsync(
             deleteRequest,
@@ -616,7 +643,7 @@ public class RemoteClusterIndicesClientTests {
             .build();
 
         ArgumentCaptor<BulkRequest> bulkRequestCaptor = ArgumentCaptor.forClass(BulkRequest.class);
-        when(mockedOpenSearchClient.bulk(bulkRequestCaptor.capture())).thenReturn(bulkResponse);
+        when(mockedOpenSearchAsyncClient.bulk(bulkRequestCaptor.capture())).thenReturn(CompletableFuture.completedFuture(bulkResponse));
 
         BulkDataObjectResponse response = sdkClient.bulkDataObjectAsync(bulkRequest, testThreadPool.executor(TEST_THREAD_POOL))
             .toCompletableFuture()
@@ -682,7 +709,7 @@ public class RemoteClusterIndicesClientTests {
             .errors(true)
             .build();
 
-        when(mockedOpenSearchClient.bulk(any(BulkRequest.class))).thenReturn(bulkResponse);
+        when(mockedOpenSearchAsyncClient.bulk(any(BulkRequest.class))).thenReturn(CompletableFuture.completedFuture(bulkResponse));
 
         BulkDataObjectResponse response = sdkClient.bulkDataObjectAsync(bulkRequest, testThreadPool.executor(TEST_THREAD_POOL))
             .toCompletableFuture()
@@ -708,11 +735,13 @@ public class RemoteClusterIndicesClientTests {
 
         BulkDataObjectRequest bulkRequest = BulkDataObjectRequest.builder().build().add(putRequest);
 
-        when(mockedOpenSearchClient.bulk(any(BulkRequest.class))).thenThrow(
-            new OpenSearchException(
-                new ErrorResponse.Builder().error(
-                    new ErrorCause.Builder().type("parse_exception").reason("Failed to parse data object in a bulk response").build()
-                ).status(RestStatus.INTERNAL_SERVER_ERROR.getStatus()).build()
+        when(mockedOpenSearchAsyncClient.bulk(any(BulkRequest.class))).thenReturn(
+            CompletableFuture.failedFuture(
+                new OpenSearchException(
+                    new ErrorResponse.Builder().error(
+                        new ErrorCause.Builder().type("parse_exception").reason("Failed to parse data object in a bulk response").build()
+                    ).status(RestStatus.INTERNAL_SERVER_ERROR.getStatus()).build()
+                )
             )
         );
 
@@ -723,6 +752,10 @@ public class RemoteClusterIndicesClientTests {
 
         CompletionException ce = assertThrows(CompletionException.class, () -> future.join());
         Throwable cause = ce.getCause();
+        assertEquals(OpenSearchStatusException.class, cause.getClass());
+        assertEquals(RestStatus.INTERNAL_SERVER_ERROR, ((OpenSearchStatusException) cause).status());
+        ce = (CompletionException) cause.getCause();
+        cause = ce.getCause();
         assertEquals(OpenSearchException.class, cause.getClass());
         assertEquals(RestStatus.INTERNAL_SERVER_ERROR.getStatus(), ((OpenSearchException) cause).status());
         assertTrue(cause.getMessage().contains("Failed to parse data object in a bulk response"));
@@ -745,8 +778,8 @@ public class RemoteClusterIndicesClientTests {
 
         ArgumentCaptor<SearchRequest> getRequestCaptor = ArgumentCaptor.forClass(SearchRequest.class);
         ArgumentCaptor<Class<Map>> mapClassCaptor = ArgumentCaptor.forClass(Class.class);
-        when(mockedOpenSearchClient.search(getRequestCaptor.capture(), mapClassCaptor.capture())).thenReturn(
-            (SearchResponse<Map>) searchResponse
+        when(mockedOpenSearchAsyncClient.search(getRequestCaptor.capture(), mapClassCaptor.capture())).thenReturn(
+            CompletableFuture.completedFuture((SearchResponse<Map>) searchResponse)
         );
 
         SearchDataObjectResponse response = sdkClient.searchDataObjectAsync(searchRequest, testThreadPool.executor(TEST_THREAD_POOL))
@@ -754,7 +787,7 @@ public class RemoteClusterIndicesClientTests {
             .join();
 
         ArgumentCaptor<SearchRequest> requestCaptor = ArgumentCaptor.forClass(SearchRequest.class);
-        verify(mockedOpenSearchClient, times(1)).search(requestCaptor.capture(), any());
+        verify(mockedOpenSearchAsyncClient, times(1)).search(requestCaptor.capture(), any());
         assertEquals(1, requestCaptor.getValue().index().size());
         assertEquals(TEST_INDEX, requestCaptor.getValue().index().get(0));
         Query query = requestCaptor.getValue().query();
@@ -784,7 +817,9 @@ public class RemoteClusterIndicesClientTests {
             .searchSourceBuilder(searchSourceBuilder)
             .build();
 
-        when(mockedOpenSearchClient.search(any(SearchRequest.class), any())).thenThrow(new UnsupportedOperationException("test"));
+        when(mockedOpenSearchAsyncClient.search(any(SearchRequest.class), any())).thenReturn(
+            CompletableFuture.failedFuture(new UnsupportedOperationException("test"))
+        );
         CompletableFuture<SearchDataObjectResponse> future = sdkClient.searchDataObjectAsync(
             searchRequest,
             testThreadPool.executor(TEST_THREAD_POOL)
@@ -792,6 +827,10 @@ public class RemoteClusterIndicesClientTests {
 
         CompletionException ce = assertThrows(CompletionException.class, () -> future.join());
         Throwable cause = ce.getCause();
+        assertEquals(OpenSearchStatusException.class, cause.getClass());
+        assertEquals(RestStatus.INTERNAL_SERVER_ERROR, ((OpenSearchStatusException) cause).status());
+        ce = (CompletionException) cause.getCause();
+        cause = ce.getCause();
         assertEquals(UnsupportedOperationException.class, cause.getClass());
         assertEquals("test", cause.getMessage());
     }
@@ -800,7 +839,7 @@ public class RemoteClusterIndicesClientTests {
     public void testSearchDataObject_NullTenantNoMultitenancy() throws IOException {
         // Tests no status exception if multitenancy not enabled
         SdkClient sdkClientNoTenant = SdkClientFactory.wrapSdkClientDelegate(
-            new RemoteClusterIndicesClient(mockedOpenSearchClient, null),
+            new RemoteClusterIndicesClient(mockedOpenSearchAsyncClient, null),
             false
         );
 
@@ -811,7 +850,9 @@ public class RemoteClusterIndicesClientTests {
             .searchSourceBuilder(searchSourceBuilder)
             .build();
 
-        when(mockedOpenSearchClient.search(any(SearchRequest.class), any())).thenThrow(new UnsupportedOperationException("test"));
+        when(mockedOpenSearchAsyncClient.search(any(SearchRequest.class), any())).thenReturn(
+            CompletableFuture.failedFuture(new UnsupportedOperationException("test"))
+        );
         CompletableFuture<SearchDataObjectResponse> future = sdkClientNoTenant.searchDataObjectAsync(
             searchRequest,
             testThreadPool.executor(TEST_THREAD_POOL)
@@ -819,6 +860,10 @@ public class RemoteClusterIndicesClientTests {
 
         CompletionException ce = assertThrows(CompletionException.class, () -> future.join());
         Throwable cause = ce.getCause();
+        assertEquals(OpenSearchStatusException.class, cause.getClass());
+        assertEquals(RestStatus.INTERNAL_SERVER_ERROR, ((OpenSearchStatusException) cause).status());
+        ce = (CompletionException) cause.getCause();
+        cause = ce.getCause();
         assertEquals(UnsupportedOperationException.class, cause.getClass());
         assertEquals("test", cause.getMessage());
     }